GrammianAngularFields.jl Documentation

GrammianAngularFields.jl Documentation

GrammianAngularFields.jl Documentation

Functions

Index

Functions

GrammianAngularFields.testme — Method.

testme(x)

Test any kind of functionality which may change at any time. This is indeed the very definition of a random function.

Examples

julia> foo = 42
42

source

GrammianAngularFields.grammify — Method.

grammify(x)

Turns a univariate 1-D array into a Grammian Angular Field sporting a 2-D matrix shape.

Examples

julia> grammify(1:4)
4×4 Array{Float64,2}:
  1.0        0.333333  -0.333333  -1.0     
  0.333333  -0.777778  -1.0       -0.333333
 -0.333333  -1.0       -0.777778   0.333333
 -1.0       -0.333333   0.333333   1.0     

source

GrammianAngularFields.plotgrammian — Method.

plotgrammian(g)

Plot a grammified univariate array.

Examples

julia> GrammianAngularFields.plotgrammian(grammify(1:10))
data: [
  "heatmap with fields type and z"
]

layout: "layout with field margin"

source

GrammianAngularFields.plotgrammian2 — Method.

plotgrammian2(x)

Examples

julia> GrammianAngularFields.plotgrammian2(sin.(1:30))
data: [
  "scatter with fields type, x, xaxis, y, and yaxis",
  "heatmap with fields type, xaxis, yaxis, and z"
]

layout: "layout with fields margin, xaxis1, xaxis2, yaxis1, and yaxis2"

source

GrammianAngularFields.normalize — Method.

normalize(x)

Transforms a univariate array into a 0 mean and unit standard deviation array.

Examples

julia> GrammianAngularFields.normalize(collect(1:5))
5-element Array{Float64,1}:
 -1.2649110640673518
 -0.6324555320336759
  0.0               
  0.6324555320336759
  1.2649110640673518

source

GrammianAngularFields.standardize — Method.

standardize(x)

Transforms a univariate array into a 0, 1 range.

source

GrammianAngularFields.standardize — Method.

standardize(x, a, b)

Transforms a univariate array into a a, b range.

source

Index

GrammianAngularFields.grammify

GrammianAngularFields.normalize

GrammianAngularFields.plotgrammian

GrammianAngularFields.plotgrammian2

GrammianAngularFields.standardize

GrammianAngularFields.standardize

GrammianAngularFields.testme